IGAPP-496: Move SnackbarContainer to containers
